Checkpoint 6: Move the Bad Guys
Two new methods, badGuyMove and charLocator. charLocator searches a given array and finds a given character in it, makes an array with the indices. badGuyMove uses the coordinates of all '%' characters and uses an updated TryMove to pick a random viable spot for them to move to. They can choose the same spot and one of them ceases to exist, not sure how to fix that yet.